Restructure remote plugin
